refactor(shared): move Tenant infrastructure to Shared
The pieces that propagate the active tenant through Doctrine —
TenantContext, the marker interfaces TenantScoped / TenantAware, the
SQL filter and its configurator, the prePersist listener, and the HTTP
request subscriber — all belong in the same shared kernel as the
aggregate they serve. Moves them out of Identity into Shared:

  - TenantContext     -> Shared/Application/TenantContext.php
  - TenantAware       -> Shared/Application/TenantAware.php
  - TenantScoped      -> Shared/Application/TenantScoped.php
  - TenantFilter             -> Shared/Infrastructure/Doctrine/Filter/
  - TenantFilterConfigurator -> Shared/Infrastructure/Doctrine/Filter/
  - TenantAssignmentListener -> Shared/Infrastructure/Doctrine/EventListener/
  - RequestTenantSubscriber  -> Shared/Infrastructure/Http/

Renames preserve git history (~99% similarity). 24 consumer files have
their `use App\Identity\Application\Tenant{Context,Aware,Scoped}` and
the Doctrine filter / listener FQCNs rewritten to point at Shared.
`config/packages/doctrine.yaml` `filters.tenant.class` now references
the new namespace.

Identity bounded context now owns User, Role, Permission, RefreshToken
and three thin Application services that genuinely belong there
(CurrentTenantProvider, RbacSeeder, RefreshTokenService) — everything
tenant-isolation related lives under Shared.

Refs #153
