+++ apps/api/src/Catalog/Application/BuiltInAssociationTypeSeeder.php
@@ -6,7 +6,7 @@ namespace App\Catalog\Application;
 
 use App\Catalog\Domain\Entity\AssociationType;
 use App\Catalog\Infrastructure\Doctrine\Repository\AssociationTypeRepository;
-use App\Identity\Application\TenantContext;
+use App\Shared\Application\TenantContext;
 use App\Shared\Domain\Tenant;
 use Doctrine\ORM\EntityManagerInterface;
