+++ apps/api/src/Catalog/Domain/Entity/AttributeGroup.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\Catalog\Domain\Entity;
 
 use App\Catalog\Infrastructure\Doctrine\Repository\AttributeGroupRepository;
-use App\Identity\Application\TenantScoped;
+use App\Shared\Application\TenantScoped;
 use App\Shared\Domain\Tenant;
 use DateTimeImmutable;
 use Doctrine\DBAL\Types\Types;
