+++ apps/api/src/Catalog/Domain/Entity/ObjectValue.php
@@ -6,7 +6,7 @@ namespace App\Catalog\Domain\Entity;
 
 use App\Catalog\Domain\Provenance;
 use App\Catalog\Infrastructure\Doctrine\Repository\ObjectValueRepository;
-use App\Identity\Application\TenantScoped;
+use App\Shared\Application\TenantScoped;
 use App\Shared\Domain\Tenant;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
