+++ apps/api/src/Identity/Application/CurrentTenantProvider.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Identity\Application;
 
+use App\Shared\Application\TenantAware;
 use App\Shared\Domain\Tenant;
 use App\Shared\Infrastructure\Doctrine\Repository\DoctrineTenantRepository;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
