+++ apps/api/tests/Functional/Asset/AssetUploaderTest.php
@@ -6,7 +6,7 @@ namespace App\Tests\Functional\Asset;
 
 use App\Asset\Application\AssetUploader;
 use App\Asset\Domain\Entity\AssetVariant;
-use App\Identity\Application\TenantContext;
+use App\Shared\Application\TenantContext;
 use App\Shared\Domain\Tenant;
 use Doctrine\ORM\EntityManagerInterface;
 use League\Flysystem\FilesystemOperator;
