+++ apps/api/tests/Unit/Identity/TenantAssignmentListenerTest.php
@@ -7,10 +7,10 @@ namespace App\Tests\Unit\Identity;
 use App\Catalog\Domain\Entity\CatalogObject;
 use App\Catalog\Domain\Entity\ObjectType;
 use App\Catalog\Domain\ObjectKind;
-use App\Identity\Application\TenantContext;
-use App\Identity\Application\TenantScoped;
-use App\Identity\Infrastructure\Doctrine\EventListener\TenantAssignmentListener;
+use App\Shared\Application\TenantContext;
+use App\Shared\Application\TenantScoped;
 use App\Shared\Domain\Tenant;
+use App\Shared\Infrastructure\Doctrine\EventListener\TenantAssignmentListener;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Event\PrePersistEventArgs;
 use LogicException;
